feat: 컴포넌트 목적 수정 및 RoundButton 적용
