+++ src/pages/TrainingDetail/components/ConfirmBottomSheet/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ConfirmBottomSheet.view";
