--- src/pages/TrainingDetail/components/ConfirmButton/ConfirmButton.scss
@@ -1,3 +1,0 @@
-.confirm-button {
-  
-}
