--- src/pages/TrainingDetail/components/ConfirmButton/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ConfirmButton.view";
